Don't use a template for nsswitch.conf
It will more than likely get into a never-ending fight with
authconfig. authconfig does not currently handle the sudoers line
though so append this if it is missing and then add or remove sss
(not ldap!) as necessary. Finally, the ldap_sudo attribute should be a
real boolean, not a string. Fixes #12.