Use ldap_tls_cacertdir rather than ldap_tls_cacert
ldap_tls_cacertdir expects a directory containing individual
certificate files ending in .0. RHEL does not install these and yet
the default attribute value was specific to RHEL. It is least
confusing to use ldap_tls_cacert on all systems and point it at the
bundle file. Both RHEL and Debian include utilities to manage this
file.