Fix libsss-sudo installation
On Debian, this is called libsss-sudo, not libsss_sudo. It is not
necessary to install it on other platforms anyway as they bundle it
with sssd or sssd-common.

Older versions of Debian and Ubuntu have libsss-sudo0 rather than
libsss-sudo. Despite the existence of this package, they annoyingly
forgot to build sudo with SSSD support. We'll install libsss-sudo0
anyway to prevent the recipe from breaking. Users may also wish to
roll their own sudo.